requirements.txt
anyio==3.6.1
attrs==22.1.0
black==22.8.0
certifi==2022.6.15
click==8.1.3
fastapi==0.81.0
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pydantic==1.10.1
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.19.0
python-dotenv==0.20.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
